Restore length limits to string columns in the database.
We utilise the difference between string and text to distinguish between short
and long pieces of textual data. It is implicitly assumed that string fields
are suited for display in areas with limited space, e.g. flash messages and
tables.

Rails 4.2 removed the length limits (see Rails' #14579.) This allows a user to
potentially enter a long piece of data, and we would put that string in the
flash message (in the guise of 'you have successfully created X'.) Since we
use the cookie session store, which has a length limit of 4KB, this would
cause an exception at runtime. That also disrupts the page flow of tables,
when the title is used to indicate the link to an object.

Since Schema Plus Validations handles the length field properly, this would
also prevent the user from submitting such long data.