Allow rendering a partial with a prefix.
This is exceptionally useful when rendering a polymorphic collection: we want to use to_partial_path, but also depending on the context we are using it, different partials might need to be used. This allows a view to specify the prefix/suffix of a partial, which will be used together with the generated partial path, to customise the partial used in a context.