ActiveRecord::Base#wait deadlock fix #1.
I'm not entirely sure if this is the problem, but one potential source of problems is when wait returns even if it is not the signal we are looking for. This ensures that the signal notified must be the signal we are waiting on before returning from the wait loop.