Define `Attachment#open`.
This encapsulates the operation of needing to read from the attachment. Currently this (dumbly) reads from the attachment by loading into memory, but this guarantees file agnosticism when using cloud storage.

In future, it might be possible to directly open a read-only stream if the attachment is on disk, but that is not currently implemented.