Move the `display: none` declaration to the toolbar element.
This guarantees that the element is hidden on non-JavaScript browsers, and that it is shown in JavaScript-enabled browsers. If this is not done, it is possible that the CSS loads after the script, and $.show will not set a `display: block` on the toolbar div. When the CSS does load, the toolbar will be hidden.