Don't apply tooltip to the Facebook like button
The current design of Coursemology will attach a tooltip to any element
with a "title" attribute. This is incompatible with the Facebook like
button. We do not want a tooltip attached to the Facebook button.

This fix is required to support any layout that wants to include the
official Facebook like button.