Fix Duplicator to support single item collections.
Also treat ActiveRecord CollectionProxy like an array.
Use ability of Duplicator's duplicate function to dup arrays in the
various `initialize_duplicate` functions.

Update duplicator spec to use the same technique in the sample
`initialize_duplicate` functions.

Update duplication service to pass in the course as an object instead of
a single element array because we want an object to be returned.