Fix Summernote image upload resize failing randomly
When the image is cached, the image loads instantly and height/width is
non-zero. However, when image is not cached, the code sees the height
and width as 0 before the image is loaded, and decides that the image
does not need to be resized.

This commit fixes the bug by waiting for the image onload event before
checking the width/height.