Fix race condition which results in unsynced submission and answer states.
Problem: There are instances of submitted submissions with answers that are still in an attempting state.
Investigation into this surfaced a race condition - given that user has 2 tabs open, the user can send 2 requests to the coursemology server: (i) Finalise submission, (ii) Submit and evaluat the answer for the current programming question being worked on.

The race condition occurs because (ii) would recreate an attempting answer after autograding is done, but is unaware that the submission could be finalised during this process. This would result in (i) having a SubJobError where the job is unable to cope with the IllegalStateError raised.

The desync in states results in the course_staff being unable to grade the answers, requring some intervention.

Solution: Two fixes are done to ensure the following -
1. Answers are not created for reattempting if the submission is submitted
2. Submissions would auto-grade all answers, including those in attempting state (which could have been created after finalising the submission, but before the autograding job has run)

The solutions proposed greatly reduces the possibility of the race condition happening, but there is a very minute possibility (in a perfect world with 0 latency) that the race condition can still occur. Hence, the submission autograding job is implemented with retries to try to catch such cases.