Display current course start time on duplication page
The current course start time is displayed and the default duplicated
course start time is set to the same time as that of the current course.
This is important because duplicated course will have start and end date
of all items shifted by the difference between the start times of the two
courses. Default start and end times of a newly created course is rounded
to the nearest hour for a cleaner UI.