Evaluate packages directly in ProgrammingEvaluationService.
Previously, the ProgrammingEvaluationService created evaluations and
waited for external evaluators to run the evaluation and update its
state.

To elimimate the polling time of these external evaluators, move actual
evaluation using docker containers into the service so the worker can
run it immediately.

No longer create ProgrammingEvaluation objects.
Use Ruby's timeout on the execute method instead of the old method of
using the custom AR extension for the ProgrammingEvaluation object.