Modify last_answer helper to return last non-current answer.
In the old scheme, the last graded answer was the 2nd last answer if the
submission was in the 'attempting' state. It's the last answer if the
submission is submitted.

Now, the last graded answer will usually be the last non-current answer.
If the submission is submitted, current_answer is the one to show the
results for.

Remove current_answer helper for autograded submissions.

Use a scope to fetch answers belonging to a question so a fresh query is
generated. This is needed to pick up the latest answer which has just
been created and evaluated.

Render the current_answer instead of the last answer for inline graded
answers.