Port over video player
This was based on the video player concept by Wudi, though there are a
few differences.

Firstly, the video player does not manage the YouTube API anymore.
Instead, react-player is used and wrapped around.

Secondly, the controls are separated out instead of being generated in
the VideoPlayer component. This separation prepares them to be used with Redux
later.

Lastly, the video slider is based off rc-slider instead of a cloned copy
of material-ui/slider. This new slider has integrated marking features
for questions later. There's also less to maintain now since it's no
longer cloned code (which might need to be kept up to date).

Signed-off-by: Shen Yichen <2007.yichen@gmail.com>