Add components
Both presentation and container components are added to the discussion
module.

If the component is too big, or if the component is a reusable
presentation component, they are split into separate files (one for
presentation and one for container). If not, they are defined in a
single file for brevity.

Signed-off-by: Shen Yichen <2007.yichen@gmail.com>