Enhance programming evaluation services to handle multiple reports files.
ProgrammingEvaluationService:
Pluralize variables to indicate that there are multiple reports.
Update the function for checking if there was an evaluation error.

ProgrammingImportService:
Combine test cases extracted from all the reports.

ProgrammingAutoGradingService:
Build test results from all the reports.
Only build a failed test result if that test case was not in any of the
reports.