Scroll to element on autoscroll
If a topic is set in the scroll state, the container will scroll to it
if autoscroll is not on.

If autoscroll is on, it'll scroll to it only if the timestamp of the
topic in the state is the same as the timestamp of the topic the
autoscroll finds. If not, the autoscroll overrides the topic set in
state.

Signed-off-by: Shen Yichen <2007.yichen@gmail.com>