Fix ActiveRecord::HasManyThroughOrderError errors
has_one :ansewr on Course::Assessment::Answer::ProgrammingAutoGrading
can be removed because the association is already defined in the parent
class Course::Assessment::Answer::AutoGrading. In fact, the presence of
the association is causing an ActiveRecord::HasManyThroughOrderError error
because its key :answer conflicts with the same key in the parent model.