Save Levels.
POST level data to server.
Remove existing levels and create new level objects with the new
thresholds.

Indicate on frontend if a threshold is 0 and do not send save request if
the user tries to key in 0 as a threshold.

Disable page elements when waiting for result of saving levels.
Notify user if saving succeeds or fails.

Put mass update levels logic from controller to model.
Keep only index and create for Levels. Remove now unneeded views and actions.