Send events to server
The video component will now have a timer initialized sends back
recorded events to the server. Once we confirm that the server has
received the events, we remove them from the state store.

The server is pinged even if there are no events. In this case we simply
want to refresh the session_end and the last_video_time in the backend,
so we have sufficiently up to date information if the user disconnects.

Signed-off-by: Shen Yichen <2007.yichen@gmail.com>