Alter columns and model for Video Session and events
Added a column to track the last known video time of the session. This
helps us figure out where the user has stopped on a window
close/disconnect.

Merged video time into a single column. Two events will be logged if we
need both a start and end.

Also added a column to make it possible for playback rate to be tracked.

Models are updated to validate new columns. We also add some new event
types to be tracked.

Signed-off-by: Shen Yichen <2007.yichen@gmail.com>