Add setting to assessment tabs for default lesson plan visibility.
Show another toggle on the lesson plan settings page to set if the item
should be selected in the filter when the lesson plan is loaded.

Pass visible setting information to the frontend and use for
initializing the filter settings.

Handle the toggling of the visible setting.

Re-initialize visibility settings when events are
created/updated/deleted.
Events might create new item types so these need to be added to the
filter.