Manually authorize and redirect video submissions
If a user is not authorized to edit a submission, then he is redirected
to the video show page so he can click through to his own submission
instead.

The rationale behind this behaviour is that since one can only watch a
video on a submission edit page, when sharing links, the submission edit
URL is often copied in place of the video URL. This error leads to dead
links in emails.

Signed-off-by: Shen Yichen <2007.yichen@gmail.com>