Implement read/unread for discussion topics
Use 'unread' gem to track read/unread comments, declared at a topic level.
'unread' tracks the last read time by a user, and compares that to the
updated_at column on discussion topics. Care must be taken to disable
updates to update_at for changes to topics.

Choice of unread gem is preferred over using a boolean flag as we are
tracking the read status of a topic with respect to one user, while the
pending flag is used by multiple staff to track.