Modify video_topics#from_user to include all authors
Current implementation of video topics only recognises the topic
creator as the 'user' of the topic. However, since any student or
staff can post to this topic, this change recognises all post
authors in the topic as a 'user' of the topic.

This scope is used in the comments tab for students to track read/
unread comments, and the change allows them to follow with the
conversation that they have participated in.