Add and hook up redux-persist
Use redux-persist to transparently store the Redux state into
localstorage.

redux-persist allows the state to be synced to local storage with
minimal logic on our end. For example, it handles the throttles and
potential race conditions of having 2 sources of truth. Finally it uses
reducer actions to repopulate the redux state from localstorage, so it
is easy to extend repopulation behaviour within any reducers by catching
the action fired.

The redux store for video is configured to persist it's state.

Signed-off-by: Shen Yichen <2007.yichen@gmail.com>