Remove UTF-8 byte order marks for CSV invitations
In certain cases, it is possible to have byte order marks (BOM)
at the start of a CSV file to indicate the type of encoding. This
happens by default when I save a CSV in excel.

This commit removes any UTF-8 BOM, or the BOM will be included within
the invitation name and be persisted in the database.