Reject invitations only if there are unconfirmed invitations
Solves #2595.

Currently all users are added through Course::UserInvitation.
There is a uniqueness constraint as part of validation, to ensure
that no invitation with the same course_id and email exists. However
this does not account for course_users who accepted the invitation
and left, and rejoin in the future.

This commit restricts the uniqueness check to unconfirmed invitations,
for us to track past confirmed invitations. This is preferred to
deleting the Course::UserInvitation object if the course_user leaves.