Modify has_many_attachments to update attachment_references
This introduces some changes to the has_many_attachments
declaration when one (or more) columns are specified.
Specifically, it allows for use with the front-end
WYSIWYG editor.

With this commit, attachable models (eg. Announcements) can
now insert images with an attachment/url referencing an
attachment_reference (with nil attachable), and the association
between attachable and attachment_reference is done just
before saving via a callback.

Also removed specs on attachable models to reduce test
duplication.