Allow video URLs to be changed w/ conditions
We allow URLs to be changed if there are no discussion topics or watch
session data associated with it.

The main issue with changing URLs is that the timestamps for topics and
sessions will become out of sync. However, if none of those records
exists, we will not have issues when we change the URL.

Signed-off-by: Shen Yichen <2007.yichen@gmail.com>