Call expression only once for Java Autograder.
The result from calling the test expression is needed twice. Once to
show the output on the frontend, and once for comparison with the
expected value.

However, the test expression might not be idempotent. This could lead to
correct looking output, but the test is considered failed, which is
confusing to the user.

Use a new class with a generic private variable to store the result of
calling the expression so it only has to be called once.