Assign skills to question assessments in Question controller.
Use question_assessment object to display nested fields for the
associated Skills. Check for nils as React forms will not even have the
key if the field is blank.

Add function to Question controller to update skills associated with a
question assessment.

Update building of new questions to add skills through the question
assessment association.