Add validate step for parsing attachment_reference ids
This solves two issues:
1) Duplication of course objects:
The column (and therefore the attachment_reference uuid) is
duplicated on course duplication. This commit detects such
cases (where attachment_reference's attachable points to a
different object) and creates a new attachment_reference.

2) Missing attachment_reference UUID:
The commit ignores the case when a provided uuid does not
exist in the database, and returns nil instead of raising
an exception.