Implement #column_to_email for has_many_attachments
This solves the problem when rich text with attachments
are sent out in emails. We cannot possibly guarantee that
email clients can make a call to Coursemology for the url
of attachments (as that requires them to sign into
Coursemology).

To solve this, the has_many_attachments declaration adds
an additional #column_to_email method for each column
declared. This converts image nodes in those columns to
the attachment URL directly.