fix(AppLoader): prevent redirection if 404
There is a risk of the page living in infinite loop if whatever request
in `AppLoader` responds 404 too. Even if we limit the redirection to
only happen when the response is `"Instance Not Found"`, there is a
risk of infinite loop if for some reasons the default instance is also
404. It's best to just throw it to `ErrorBoundary` and stop rendering.
The page will show a 404 message anyway.